+++ analytics/urls.py
@@ -5,11 +5,6 @@ from django.conf.urls import include
 from django.urls import path
 from django.urls.resolvers import URLPattern, URLResolver
 
-from analytics.views.installation_activity import (
-    get_installation_activity,
-    get_integrations_activity,
-)
-from analytics.views.realm_activity import get_realm_activity
 from analytics.views.stats import (
     get_chart_data,
     get_chart_data_for_installation,
@@ -19,17 +14,10 @@ from analytics.views.stats import (
     stats_for_installation,
     stats_for_realm,
 )
-from analytics.views.support import support
-from analytics.views.user_activity import get_user_activity
 from zerver.lib.rest import rest_path
 
 i18n_urlpatterns: List[Union[URLPattern, URLResolver]] = [
     # Server admin (user_profile.is_staff) visible stats pages
-    path("activity", get_installation_activity),
-    path("activity/integrations", get_integrations_activity),
-    path("activity/support", support, name="support"),
-    path("realm_activity/<realm_str>/", get_realm_activity),
-    path("user_activity/<user_profile_id>/", get_user_activity),
     path("stats/realm/<realm_str>/", stats_for_realm),
     path("stats/installation", stats_for_installation),
     # User-visible stats page
@@ -37,18 +25,14 @@ i18n_urlpatterns: List[Union[URLPattern, URLResolver]] = [
 ]
 
 if settings.ZILENCER_ENABLED:
-    from analytics.views.remote_activity import get_remote_server_activity
     from analytics.views.stats import stats_for_remote_installation, stats_for_remote_realm
-    from analytics.views.support import remote_servers_support
 
     i18n_urlpatterns += [
-        path("activity/remote", get_remote_server_activity),
         path("stats/remote/<int:remote_server_id>/installation", stats_for_remote_installation),
         path(
             "stats/remote/<int:remote_server_id>/realm/<int:remote_realm_id>/",
             stats_for_remote_realm,
         ),
-        path("activity/remote/support", remote_servers_support, name="remote_servers_support"),
     ]
 
 # These endpoints are a part of the API (V1), which uses:
