+++ corporate/urls.py
@@ -21,6 +21,10 @@ from corporate.views.event_status import (
     remote_server_event_status,
     remote_server_event_status_page,
 )
+from corporate.views.installation_activity import (
+    get_installation_activity,
+    get_integrations_activity,
+)
 from corporate.views.portico import (
     app_download_link_redirect,
     apps_view,
@@ -32,6 +36,8 @@ from corporate.views.portico import (
     remote_server_plans_page,
     team_view,
 )
+from corporate.views.realm_activity import get_realm_activity
+from corporate.views.remote_activity import get_remote_server_activity
 from corporate.views.remote_billing_page import (
     remote_billing_legacy_server_confirm_login,
     remote_billing_legacy_server_from_login_confirmation_link,
@@ -56,7 +62,7 @@ from corporate.views.sponsorship import (
     sponsorship,
     sponsorship_page,
 )
-from corporate.views.support import support_request
+from corporate.views.support import remote_servers_support, support, support_request
 from corporate.views.upgrade import (
     remote_realm_upgrade,
     remote_realm_upgrade_page,
@@ -65,6 +71,7 @@ from corporate.views.upgrade import (
     upgrade,
     upgrade_page,
 )
+from corporate.views.user_activity import get_user_activity
 from corporate.views.webhook import stripe_webhook
 from zerver.lib.rest import rest_path
 from zerver.lib.url_redirects import LANDING_PAGE_REDIRECTS
@@ -81,6 +88,14 @@ i18n_urlpatterns: Any = [
     path("support/", support_request),
     path("billing/event_status/", event_status_page, name="event_status_page"),
     path("stripe/webhook/", stripe_webhook, name="stripe_webhook"),
+    # Server admin (user_profile.is_staff) visible stats pages
+    path("activity", get_installation_activity),
+    path("activity/integrations", get_integrations_activity),
+    path("activity/support", support, name="support"),
+    path("realm_activity/<realm_str>/", get_realm_activity),
+    path("user_activity/<user_profile_id>/", get_user_activity),
+    path("activity/remote", get_remote_server_activity),
+    path("activity/remote/support", remote_servers_support, name="remote_servers_support"),
 ]
 
 v1_api_and_json_patterns = [
